+++ Runtime/HMF/IEntity.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+
+namespace HMF.HMFUtilities
+{
+    public interface IEntity
+    {
+        Transform transform { get; }
+    }
+}
+