+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Changelog
 
-## Features V1.0.0 
-Basic setup of the package.
+## [1.0.0]
+### Added
+- File structure and package.json
 
-## Added
-File structure and package.json
+### Removed
+- All unnecessary meta files
+
+## [1.0.1] - 2020.10.13.
+### Added
+- IEntity 
+- Command for the Command Pattern
+- CommandProcessor that handles all Commands
+
+
+## [1.0.2] - 2020.10.19.
+### Added
+- Singleton Pattern
+- Generic Singleton that derives from MonoBehaviour both persistent and non